[fix](be) Harden and optimize native Parquet decoding
### What problem does this PR solve?

Issue Number: None

Related PR: #65674

Problem Summary: FileScannerV2 native Parquet scans could mis-handle malformed encoding sizes and large skips, reject valid nested MAP shapes, miss shifted STRUCT siblings, and lose legacy timestamp timezone semantics. The scan path also repeated dictionary conversion, selection bitmap construction, metadata adaptation, and predicate work, while profile counters were not always published or attributed at a diagnosable layer. Harden decoder bounds and exact-size checks, preserve complex-type and timestamp invariants, reuse persistent typed state and scratch with hysteresis, adapt predicate ordering and staged prefetch, and complete the Scanner-to-IO profile hierarchy and documentation.

### Release note

Improve FileScannerV2 Parquet correctness, performance, memory stability, and scan-profile attribution.

### Check List (For Author)

- Test: Unit Test and Regression test
    - 309 Parquet BE unit tests; 92 scanner/table-reader unit tests; 39 focused review/performance/profile unit tests; 3 related local TVF regression suites
    - BE and FE builds
    - C++ format check and diff check
- Behavior changed: Yes (malformed Parquet input now fails safely; valid nested MAP/STRUCT and legacy timestamp semantics are preserved; scan scheduling and Profile attribution are optimized)
- Does this need documentation: Yes (review guide and FileScannerV2 design documents are updated in this PR)

+++ be/src/format_v2/parquet/reader/native/byte_array_plain_decoder.cpp
@@ -26,7 +26,7 @@
 
 namespace doris::format::parquet::native {
 namespace {
-Status read_length(const Slice* data, uint32_t* offset, uint32_t* length) {
+Status read_length(const Slice* data, size_t* offset, uint32_t* length) {
     if (UNLIKELY(*offset > data->size || data->size - *offset < sizeof(uint32_t))) {
         return Status::IOError("Can't read byte array length from plain decoder");
     }
